+++ src/domyoselliptical.cpp
@@ -7,7 +7,7 @@
 #include <QBluetoothLocalDevice>
 #include <QSettings>
 
-domyoselliptical::domyoselliptical(bool noWriteResistance, bool noHeartService, bool testResistance, uint8_t bikeResistanceOffset, uint8_t bikeResistanceGain)
+domyoselliptical::domyoselliptical(bool noWriteResistance, bool noHeartService, bool testResistance, uint8_t bikeResistanceOffset, double bikeResistanceGain)
 {
     m_watt.setType(metric::METRIC_WATT);
     refresh = new QTimer(this);
