+++ src/echelonconnectsport.h
@@ -38,7 +38,7 @@ class echelonconnectsport : public bike
 {
     Q_OBJECT
 public:
-    echelonconnectsport(bool noWriteResistance, bool noHeartService, uint8_t bikeResistanceOffset, uint8_t bikeResistanceGain);
+    echelonconnectsport(bool noWriteResistance, bool noHeartService, uint8_t bikeResistanceOffset, double bikeResistanceGain);
     bool connected();
 
     void* VirtualBike();
@@ -63,7 +63,7 @@ private:
     QLowEnergyCharacteristic gattNotify2Characteristic;
 
     uint8_t bikeResistanceOffset = 4;
-    uint8_t bikeResistanceGain = 1;
+    double bikeResistanceGain = 1.0;
     uint8_t counterPoll = 1;
     QDateTime lastTimeUpdate;
     bool firstUpdate = true;
